+++ ui/gfx/animation/linear_animation.cc
@@ -46,7 +46,7 @@ double LinearAnimation::GetCurrentValue() const {
 }
 
 void LinearAnimation::SetCurrentValue(double new_value) {
-  new_value = std::max(0.0, std::min(1.0, new_value));
+  new_value = std::max((double)0.0, std::min((double)1.0, new_value));
   base::TimeDelta time_delta = base::TimeDelta::FromMicroseconds(
       duration_.InMicroseconds() * (new_value - state_));
   SetStartTime(start_time() - time_delta);
